Move Semigroup-related instances for Down and Symbol
This will be necessary to avoid import cycles with `Data.Ord.Singletons` in a
subsequent commit.
